requirements.txt
# Core ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
xgboost==2.1.1
prophet==1.1.6

# MLOps (MLflow 3.x)
mlflow>=3.0.0

# Data versioning
dvc==3.55.2

# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2

# Monitoring
prometheus-client==0.21.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
matplotlib==3.9.2
plotly==5.24.1
scipy==1.14.1
joblib==1.4.2